fix: ensures process is not undefined before logging to console
In some instances, the global `process` object is undefined.
If it is undefined, an app using `react-context-responsive` crashes:
> Uncaught ReferenceError: process is not defined

One example scenario where this can occur is when using webpack 5. This is because as of webpack 5:
"...no longer include a polyfill for this Node.js variable. Avoid using it in the frontend code."
For more info, see "Level 5: Runtime Errors" here:
https://webpack.js.org/migrate/5/#run-a-single-build-and-follow-advises

This commit wraps the `process` conditional check in a `typeof` comparison,
to ensure it's not `undefined`.